Move normalizeImportOrderOption
