+++ src/utils/normalize-plugin-options.ts
